style: Format with black 25.x to match CI
Formatting-only; 8 of 9 files drifted in the windows-installer merge.

+++ ndviewer_light/core.py
@@ -3535,9 +3535,7 @@ class LightweightViewer(QWidget):
                 if not filepath:
                     return np.zeros((height, width), dtype=plane_dtype)
                 try:
-                    return decode_image_plane(filepath).astype(
-                        plane_dtype, copy=False
-                    )
+                    return decode_image_plane(filepath).astype(plane_dtype, copy=False)
                 except Exception as e:
                     logger.debug("Failed to load plane %s: %s", filepath, e)
                 return np.zeros((height, width), dtype=plane_dtype)
